requirements.txt
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0
nltk>=3.8.0
spacy>=3.6.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
langchain>=0.0.300
langchain-openai>=0.0.1
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6 